[Finishes #1639333451] Move test_data from tests directory
